refactor: move validate-env-names.cjs into archunit-tests package
The script is only used by env-vars-naming.archunit.test.ts and has no
other consumers. Moving it alongside the test removes the cross-package
deep import and keeps related code co-located.

- build-pipeline/scripts/validate-env-names.cjs: deleted
- packages/ocom/archunit-tests/src/validate-env-names.cjs: new location
- env-vars-naming.archunit.test.ts: import updated to './validate-env-names.cjs'
- .github/copilot-instructions.md: fix stale path reference
- validatedBy field updated to reflect canonical path
- fix Biome lint errors now that file is inside a linted package
  (noAssignInExpressions, noUnusedVariables)
