+++ .github/copilot-instructions.md
@@ -162,7 +162,7 @@ Frontend environment variables follow a strict naming convention to enable pipel
 - Portal-specific: VITE_APP_<PORTAL_NAME>_<ENV_VAR_NAME> (e.g., VITE_APP_UI_COMMUNITY_B2C_CLIENTID)
 - Shared/common: VITE_COMMON_<ENV_NAME> (e.g., VITE_COMMON_API_ENDPOINT)
 
-See apps/docs/docs/decisions/0031-ui-env-vars.md for full details and examples. Agents and automated tooling MUST validate generated or modified VITE_* names using scripts/validate-env-names.js and must not introduce non-conforming names. The canonical enforcement test lives in packages/ocom/archunit-tests/src/env-vars-naming.archunit.test.ts.
+See apps/docs/docs/decisions/0031-ui-env-vars.md for full details and examples. Agents and automated tooling MUST validate generated or modified VITE_* names using `packages/ocom/archunit-tests/src/validate-env-names.cjs` and must not introduce non-conforming names. The canonical enforcement test lives in packages/ocom/archunit-tests/src/env-vars-naming.archunit.test.ts.
 
 Example guidance for agents:
 - If a variable is specific to the Community portal, name it VITE_APP_UI_COMMUNITY_<NAME>
